move get tenant to file
