+++ src/services/lease-service/priority-list-service.ts
@@ -6,7 +6,6 @@
 
 import {
   Applicant,
-  Contact,
   DetailedApplicant,
   Lease,
   Listing,
@@ -23,156 +22,8 @@ import {
   getLeasesForContactCode,
   getResidentialAreaByRentalPropertyId,
 } from './adapters/xpand/tenant-lease-adapter'
-import { getEstateCodeFromXpandByRentalObjectCode } from './adapters/xpand/estate-code-adapter'
 
 export type AdapterResult<T, E> = { ok: true; data: T } | { ok: false; err: E }
-
-type GetTenantError =
-  | 'get-contact'
-  | 'contact-not-found'
-  | 'contact-not-tenant'
-  | 'get-waiting-lists'
-  | 'waiting-list-internal-parking-space-not-found'
-  | 'get-contact-leases'
-  | 'contact-leases-not-found'
-  | 'get-residential-area'
-  | 'housing-contracts-not-found'
-  | 'get-lease-property-info'
-
-export type Tenant = Omit<Contact, 'leases'> & {
-  isTenant: true
-  queuePoints: number
-  currentHousingContract?: Lease
-  upcomingHousingContract?: Lease
-  parkingSpaceContracts?: Lease[]
-}
-
-export async function getTenant(params: {
-  contactCode: string
-}): Promise<AdapterResult<Tenant, GetTenantError>> {
-  const contact = await getContactByContactCode(params.contactCode, 'false')
-  if (!contact.ok) {
-    return { ok: false, err: 'get-contact' }
-  }
-
-  if (!contact.data) {
-    return { ok: false, err: 'contact-not-found' }
-  }
-
-  if (contact.data.isTenant !== true) {
-    return { ok: false, err: 'contact-not-tenant' }
-  }
-
-  const waitingList = await getWaitingList(
-    contact.data.nationalRegistrationNumber
-  )
-
-  if (!waitingList.ok) {
-    return { ok: false, err: 'get-waiting-lists' }
-  }
-
-  const waitingListForInternalParkingSpace =
-    parseWaitingListForInternalParkingSpace(waitingList.data)
-
-  if (!waitingListForInternalParkingSpace) {
-    return {
-      ok: false,
-      err: 'waiting-list-internal-parking-space-not-found',
-    }
-  }
-
-  const leases = await getLeasesForContactCode(
-    contact.data.contactCode,
-    'true',
-    undefined
-  )
-
-  if (!leases.ok) {
-    return {
-      ok: false,
-      err: 'get-contact-leases',
-    }
-  }
-
-  if (!leases.data.length) {
-    return {
-      ok: false,
-      err: 'contact-leases-not-found',
-    }
-  }
-
-  const activeAndUpcomingLeases = leases.data.filter(isLeaseActiveOrUpcoming)
-
-  const leasesWithResidentialArea = await Promise.all(
-    activeAndUpcomingLeases.map(async (lease) => {
-      const residentialArea = await getResidentialAreaByRentalPropertyId(
-        lease.rentalPropertyId
-      )
-
-      if (!residentialArea.ok) {
-        throw new Error('Err getting residential area')
-      }
-
-      return {
-        ...lease,
-        residentialArea: residentialArea.data,
-      }
-    })
-  )
-    .then((data) => ({ ok: true, data }) as const)
-    .catch((err) => ({ ok: false, err }) as const)
-
-  if (!leasesWithResidentialArea.ok) {
-    return { ok: false, err: 'get-residential-area' }
-  }
-
-  const housingContracts = parseLeasesForHousingContracts(
-    leasesWithResidentialArea.data
-  )
-
-  if (!housingContracts) {
-    return { ok: false, err: 'housing-contracts-not-found' }
-  }
-
-  const [currentHousingContract, upcomingHousingContract] = housingContracts
-
-  if (!currentHousingContract && !upcomingHousingContract) {
-    return { ok: false, err: 'housing-contracts-not-found' }
-  }
-
-  const leasesWithPropertyType = await Promise.all(
-    leasesWithResidentialArea.data.map(async (l) => {
-      const type = await getEstateCodeFromXpandByRentalObjectCode(
-        l.rentalPropertyId
-      ).then((v) => v?.type)
-
-      return { ...l, propertyType: type }
-    })
-  )
-    .then((data) => ({ ok: true, data }) as const)
-    .catch((err) => ({ ok: false, err }) as const)
-
-  if (!leasesWithPropertyType.ok) {
-    return { ok: false, err: 'get-lease-property-info' }
-  }
-
-  const parkingSpaceContracts = parseLeasesForParkingSpaces(
-    leasesWithPropertyType.data
-  )
-
-  return {
-    ok: true,
-    data: {
-      ...contact.data,
-      isTenant: contact.data.isTenant,
-      queuePoints: waitingListForInternalParkingSpace.queuePoints,
-      address: contact.data.address,
-      currentHousingContract,
-      upcomingHousingContract,
-      parkingSpaceContracts,
-    },
-  }
-}
 
 type GetDetailedApplicantError =
   | 'get-contact'
